Fix player status not being updated during setup.
Fixes #1209.
